fix(server): inline api loader for node runtime
