+++ apps/server/api/[[...path]].ts
@@ -1,4 +1,16 @@
-import { getAppInstance } from "./load-app";
+import type { createApp as CreateAppFactory } from "../src/app";
+
+type AppFactory = typeof CreateAppFactory;
+type AppInstance = ReturnType<AppFactory>;
+
+let appPromise: Promise<AppInstance> | null = null;
+
+const preferCompiledBundle = process.env.VERCEL === "1" || process.env.VERCEL === "true";
+const compiledBundleSpecifier = "../dist/index.js";
+const bundleMissingExportMessage =
+	"[server] Compiled bundle missing createApp export; ensure dist/index.js re-exports createApp";
+const bundleLoadFailureMessage =
+	"[server] Unable to load compiled app bundle from dist/index.js; rebuild the server package";
 
 export const config = {
 	runtime: "nodejs",
@@ -6,13 +18,14 @@ export const config = {
 
 export default {
 	async fetch(request: Request) {
-		let app;
+		let app: AppInstance;
 		try {
 			app = await getAppInstance();
 		} catch (error) {
 			console.error("[server] failed to load Elysia app", error);
 			return new Response("Internal Server Error", { status: 500 });
 		}
+
 		const originalUrl = new URL(request.url);
 		const targetPath = normalizeApiPath(originalUrl.pathname);
 		const targetUrl = new URL(targetPath + originalUrl.search, originalUrl.origin);
@@ -27,7 +40,11 @@ export default {
 				error,
 			});
 			const message =
-				error instanceof Error ? error.stack || error.message : typeof error === "string" ? error : "Internal Server Error";
+				error instanceof Error
+					? error.stack || error.message
+					: typeof error === "string"
+					? error
+					: "Internal Server Error";
 			return new Response(message, { status: 500 });
 		}
 	},
@@ -43,4 +60,38 @@ function normalizeApiPath(pathname: string) {
 	if (normalized === "") normalized = "/";
 	if (!normalized.startsWith("/")) normalized = `/${normalized}`;
 	return normalized;
+}
+
+async function loadFactory(): Promise<AppFactory> {
+	if (preferCompiledBundle) {
+		try {
+			const module = await import(compiledBundleSpecifier);
+			const createApp = (module as { createApp?: AppFactory }).createApp;
+			if (typeof createApp === "function") {
+				return createApp;
+			}
+			throw new Error(bundleMissingExportMessage);
+		} catch (error) {
+			console.error("[server] Failed to load compiled app bundle", error);
+			const cause = error instanceof Error ? error : undefined;
+			throw cause ? new Error(bundleLoadFailureMessage, { cause }) : new Error(bundleLoadFailureMessage);
+		}
+	}
+
+	if (preferCompiledBundle) {
+		console.warn(
+			"[server] Compiled bundle unavailable; falling back to TypeScript source app module."
+		);
+	} else {
+		console.debug("[server] Using TypeScript source app module (development).");
+	}
+	const module = await import("../src/app");
+	return module.createApp;
+}
+
+export async function getAppInstance(): Promise<AppInstance> {
+	if (!appPromise) {
+		appPromise = loadFactory().then((factory) => factory());
+	}
+	return appPromise;
 }
